Refactor of cartActions and types completed.
